backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
google-generativeai
werkzeug
python-dotenv